# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Receita da Boa - Social Recipe Network API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LikeResponse {
  liked: boolean;
  likeCount: number;
}

export interface FollowResponse {
  following: boolean;
}

export type UserPapel = (typeof UserPapel)[keyof typeof UserPapel];

export const UserPapel = {
  usuario: "usuario",
  adm: "adm",
} as const;

export interface User {
  id: number;
  nome: string;
  email: string;
  papel: UserPapel;
  isBanned: boolean;
}

export type UserProfilePapel =
  (typeof UserProfilePapel)[keyof typeof UserProfilePapel];

export const UserProfilePapel = {
  usuario: "usuario",
  adm: "adm",
} as const;

export interface UserProfile {
  id: number;
  nome: string;
  email: string;
  papel: UserProfilePapel;
  isBanned: boolean;
  followersCount: number;
  followingCount: number;
  receitasCount: number;
  isFollowing: boolean;
}

export interface Categoria {
  id: number;
  nome: string;
}

export interface Receita {
  id: number;
  titulo: string;
  descricao: string;
  ingredientes: string;
  instrucoes: string;
  urlImagem?: string | null;
  categoriaId?: number | null;
  autorId: number;
  isReported: boolean;
  likeCount: number;
  isLiked: boolean;
  isFavorited: boolean;
  categoria?: Categoria | null;
  autor?: User | null;
  createdAt: string;
}

export interface ReportedReceita {
  id: number;
  titulo: string;
  autorId: number;
  isReported: boolean;
  reportCount: number;
  autor?: User | null;
  categoria?: Categoria | null;
}

export interface CreateReceitaInput {
  titulo: string;
  descricao: string;
  ingredientes: string;
  instrucoes: string;
  urlImagem?: string | null;
  categoriaId?: number | null;
}

export interface CreateCategoriaInput {
  nome: string;
}

export interface FavoritoInput {
  receitaId: number;
}

export interface ReportInput {
  receitaId: number;
  motivo: string;
}

export type GetReceitasParams = {
  feed?: GetReceitasFeed;
  search?: string;
  categoriaId?: number;
};

export type GetReceitasFeed =
  (typeof GetReceitasFeed)[keyof typeof GetReceitasFeed];

export const GetReceitasFeed = {
  recentes: "recentes",
  seguindo: "seguindo",
  populares: "populares",
} as const;