# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Receita da Boa - Social Recipe Network API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: receitas
  - name: categorias
  - name: favoritos
  - name: likes
  - name: follows
  - name: comments
  - name: reports
  - name: admin
  - name: usuarios

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /receitas:
    get:
      operationId: getReceitas
      tags: [receitas]
      summary: List recipes
      parameters:
        - name: feed
          in: query
          schema:
            type: string
            enum: [recentes, seguindo, populares]
        - name: search
          in: query
          schema:
            type: string
        - name: categoriaId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receita"
    post:
      operationId: createReceita
      tags: [receitas]
      summary: Create a recipe (any logged-in user)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceitaInput"
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receita"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /receitas/{id}:
    get:
      operationId: getReceita
      tags: [receitas]
      summary: Get recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receita"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateReceita
      tags: [receitas]
      summary: Update own recipe
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceitaInput"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receita"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteReceita
      tags: [receitas]
      summary: Delete own recipe
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categorias:
    get:
      operationId: getCategorias
      tags: [categorias]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Categoria"
    post:
      operationId: createCategoria
      tags: [categorias]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoriaInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"

  /favoritos:
    get:
      operationId: getFavoritos
      tags: [favoritos]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorited recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receita"
    post:
      operationId: addFavorito
      tags: [favoritos]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoritoInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /favoritos/{receitaId}:
    delete:
      operationId: removeFavorito
      tags: [favoritos]
      security:
        - bearerAuth: []
      parameters:
        - name: receitaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /likes/{receitaId}:
    post:
      operationId: likeReceita
      tags: [likes]
      summary: Like a recipe
      security:
        - bearerAuth: []
      parameters:
        - name: receitaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /follows/{userId}:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow or unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
        "400":
          description: Cannot follow yourself
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports:
    post:
      operationId: reportReceita
      tags: [reports]
      summary: Report a recipe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: Get all reported recipes (ADM only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reported recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportedReceita"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/receitas/{id}:
    delete:
      operationId: adminDeleteReceita
      tags: [admin]
      summary: Admin delete any recipe
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/usuarios/{id}/ban:
    post:
      operationId: banUsuario
      tags: [admin]
      summary: Ban a user (ADM only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /usuarios/{id}:
    get:
      operationId: getUsuario
      tags: [usuarios]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required: [liked, likeCount]

    FollowResponse:
      type: object
      properties:
        following:
          type: boolean
      required: [following]

    User:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        papel:
          type: string
          enum: [usuario, adm]
        isBanned:
          type: boolean
      required: [id, nome, email, papel, isBanned]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        papel:
          type: string
          enum: [usuario, adm]
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        receitasCount:
          type: integer
        isFollowing:
          type: boolean
      required: [id, nome, email, papel, isBanned, followersCount, followingCount, receitasCount, isFollowing]

    Categoria:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
      required: [id, nome]

    Receita:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descricao:
          type: string
        ingredientes:
          type: string
        instrucoes:
          type: string
        urlImagem:
          type: string
          nullable: true
        categoriaId:
          type: integer
          nullable: true
        autorId:
          type: integer
        isReported:
          type: boolean
        likeCount:
          type: integer
        isLiked:
          type: boolean
        isFavorited:
          type: boolean
        categoria:
          $ref: "#/components/schemas/Categoria"
          nullable: true
        autor:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required: [id, titulo, descricao, ingredientes, instrucoes, autorId, isReported, likeCount, isLiked, isFavorited, createdAt]

    ReportedReceita:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        autorId:
          type: integer
        isReported:
          type: boolean
        reportCount:
          type: integer
        autor:
          $ref: "#/components/schemas/User"
          nullable: true
        categoria:
          $ref: "#/components/schemas/Categoria"
          nullable: true
      required: [id, titulo, autorId, isReported, reportCount]

    CreateReceitaInput:
      type: object
      properties:
        titulo:
          type: string
        descricao:
          type: string
        ingredientes:
          type: string
        instrucoes:
          type: string
        urlImagem:
          type: string
          nullable: true
        categoriaId:
          type: integer
          nullable: true
      required: [titulo, descricao, ingredientes, instrucoes]

    CreateCategoriaInput:
      type: object
      properties:
        nome:
          type: string
      required: [nome]

    FavoritoInput:
      type: object
      properties:
        receitaId:
          type: integer
      required: [receitaId]

    ReportInput:
      type: object
      properties:
        receitaId:
          type: integer
        motivo:
          type: string
      required: [receitaId, motivo]