# lib/api-zod/src/generated/types/likeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Receita da Boa - Social Recipe Network API
 * OpenAPI spec version: 0.2.0
 */

export interface LikeResponse {
  liked: boolean;
  likeCount: number;
}