# lib/api-zod/src/generated/types/reportedReceita.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Receita da Boa - Social Recipe Network API
 * OpenAPI spec version: 0.2.0
 */
import type { Categoria } from "./categoria";
import type { User } from "./user";

export interface ReportedReceita {
  id: number;
  titulo: string;
  autorId: number;
  isReported: boolean;
  reportCount: number;
  autor?: User | null;
  categoria?: Categoria | null;
}