# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Receita da Boa - Social Recipe Network API
 * OpenAPI spec version: 0.2.0
 */
import type { UserProfilePapel } from "./userProfilePapel";

export interface UserProfile {
  id: number;
  nome: string;
  email: string;
  papel: UserProfilePapel;
  isBanned: boolean;
  followersCount: number;
  followingCount: number;
  receitasCount: number;
  isFollowing: boolean;
}